% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_S2.R
\name{arcsin}
\alias{arcsin}
\title{Numerically more stable version of the arcsine of the modulus of the dot product of x and y.}
\usage{
arcsin(x, y)
}
\arguments{
\item{x}{A point on the 2-sphere.}

\item{y}{A point on the 2-sphere.}
}
\description{
Numerically more stable version of the arcsine of the modulus of the dot product of x and y.
}
\author{
Johannes Wieditz
}
