% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_S1.R
\name{branch.circle}
\alias{branch.circle}
\title{Branch computes a subdivision of a triangle}
\usage{
branch.circle(interval, init = INIT, mid = MID)
}
\arguments{
\item{interval}{A subinterval of [-pi, pi].}

\item{init}{Generates from SET a list with the following attributes
- size
- upper bound (ub)
- lower bound (lb)}

\item{mid}{A rule to compute the midpoint of the interval.}
}
\value{
A list of triangles generated by INIT.
}
\description{
Branch computes a subdivision of a triangle
}
\author{
Johannes Wieditz
}
