% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_S2.R
\name{branch.sphere}
\alias{branch.sphere}
\title{Branch computes a subdivision of a triangle.}
\usage{
branch.sphere(D, INIT, ...)
}
\arguments{
\item{D}{A triangle.}

\item{INIT}{generates from SET a list with the following attributes
- size
- upper bound (ub)
- lower bound (lb)}
}
\value{
A list of triangles generated by INIT.
}
\description{
Branch computes a subdivision of a triangle.
}
\author{
Johannes Wieditz
}
