% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BranchAndBoundWrapper.R
\name{BranchAndBound}
\alias{BranchAndBound}
\title{Branch and Bound  - Wrapper for branch and bound algorithms for computation of an (eps,delta)-approximation.}
\usage{
BranchAndBound(
  SET,
  INIT,
  BRANCH,
  MID,
  SIZE,
  p = 2,
  eps = 0.001,
  delta = 0.1,
  env = parent.frame(),
  eta = 1e-08,
  dimension,
  ...
)
}
\arguments{
\item{SET}{contains a list of points describing an initial triangulation/ subdivision of the manifold}

\item{INIT}{generates from SET a list with the following attributes
- size
- upper bound (ub)
- lower bound (lb)}

\item{BRANCH}{a function which subdivides the passed set (e.g. a triangle)}

\item{MID}{determines the midpoint of a passed set/ triangle}

\item{p}{power of the distance in the Fréchet function (cf. Fréchet-p-mean)}

\item{eps}{tolerance parameter for the function value (eps-minimality). Default is eps = 1e-3.}

\item{delta}{maximal size of the triangulation. Default is delta = 1e-1.}

\item{eta}{tolerance parameter for numerical evaluation of inequalities}

\item{dimension}{dimension of the considered sphere (e.g. 1 for the circle)}
}
\description{
Branch and Bound  - Wrapper for branch and bound algorithms for computation of an (eps,delta)-approximation.
}
\author{
Johannes Wieditz
}
