% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_S1.R
\name{darc.circle}
\alias{darc.circle}
\title{The arc length metric on the circle.}
\usage{
darc.circle(x, y)
}
\arguments{
\item{x}{The angle on a point on the unit circle.}

\item{y}{The angle on a point on the unit circle.}
}
\description{
The arc length metric on the circle.
}
\author{
Johannes Wieditz
}
