% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_S2.R
\name{darc.sphere}
\alias{darc.sphere}
\title{The arc length metric on the 2-sphere.}
\usage{
darc.sphere(x, y)
}
\arguments{
\item{x}{A point on the 2-sphere.}

\item{y}{A point on the 2-sphere.}
}
\description{
The arc length metric on the 2-sphere.
}
\author{
Johannes Wieditz
}
