% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{delete}
\alias{delete}
\title{Delete an element from a list.}
\usage{
delete(L, x)
}
\arguments{
\item{L}{A list.}

\item{x}{An element to be deleted from the list. If x is not contained in the list, then the output is again the list passed.}
}
\description{
Delete an element from a list.
}
\author{
Johannes Wieditz
}
