% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_S2.R
\name{draw.eps.delta.approximation}
\alias{draw.eps.delta.approximation}
\title{Plot the computed (eps, delta)-approximation on a 3D object.}
\usage{
draw.eps.delta.approximation(A, data)
}
\arguments{
\item{A}{The computed (eps, delta)-approximation.}

\item{data}{The input data as list of cartesian coordinates.}
}
\description{
Plot the computed (eps, delta)-approximation on a 3D object.
}
\author{
Johannes Wieditz
}
