% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{Fnhat}
\alias{Fnhat}
\title{The Fréchet-p-sample function.}
\usage{
Fnhat(m, d, ...)
}
\arguments{
\item{m}{Point at which the Fréchet function is evaluated.}

\item{d}{Metric on the considered space.}
}
\value{
Value of the Fréchet function in m.
}
\description{
The Fréchet-p-sample function.
}
\author{
Johannes Wieditz
}
