% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_S2.R
\name{get.angles.geo}
\alias{get.angles.geo}
\title{Compute geographical spherical coordinates of a 3D vector given in cartesian coordinates.}
\usage{
get.angles.geo(y)
}
\arguments{
\item{y}{A 3D vector.}
}
\description{
Compute geographical spherical coordinates of a 3D vector given in cartesian coordinates.
}
\author{
Johannes Wieditz
}
