% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{get.max.index}
\alias{get.max.index}
\title{Determine the index of the first element of the list with the highest element of the attribut entry.}
\usage{
get.max.index(L, entry)
}
\arguments{
\item{L}{A list.}

\item{entry}{An attribute of an element of the list.}
}
\description{
Determine the index of the first element of the list with the highest element of the attribut entry.
}
\author{
Johannes Wieditz
}
