% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_S2.R
\name{get.vector}
\alias{get.vector}
\title{Compute the cartesian coordinates of a 3D vector given in spherical coordinates.}
\usage{
get.vector(theta = 0, phi = 0)
}
\arguments{
\item{y}{A 3D vector.}
}
\description{
Compute the cartesian coordinates of a 3D vector given in spherical coordinates.
}
\author{
Johannes Wieditz
}
