% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_S2.R
\name{\%x\%}
\alias{\%x\%}
\title{The crossproduct of the 3D vectors.}
\usage{
a \%x\% b
}
\arguments{
\item{a}{A 3-dimensional vector.}

\item{b}{A 3-dimensional vector.}
}
\description{
The crossproduct of the 3D vectors.
}
\author{
Johannes Wieditz
}
