% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_S1.R
\name{init.3ineq.circle}
\alias{init.3ineq.circle}
\title{Formats the passed set into a data format which can be processed by the branch and bound algorithm.}
\usage{
init.3ineq.circle(set, ...)
}
\arguments{
\item{set}{A list of intervals.}
}
\description{
Formats the passed set into a data format which can be processed by the branch and bound algorithm.
}
\author{
Johannes Wieditz
}
