% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_S2.R
\name{init.3ineq.sphere}
\alias{init.3ineq.sphere}
\title{Initialise a list with set entries and lower bound using the triangle inequality.}
\usage{
init.3ineq.sphere(set, ...)
}
\arguments{
\item{set}{A list of elements to initialise.}
}
\description{
Initialise a list with set entries and lower bound using the triangle inequality.
}
\author{
Johannes Wieditz
}
