% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_S1.R
\name{init.acc.circle}
\alias{init.acc.circle}
\title{Formats the passed set into a data format which can be processed by the branch and bound algorithm.}
\usage{
init.acc.circle(set, ...)
}
\arguments{
\item{set}{A list of intervals.}
}
\description{
Formats the passed set into a data format which can be processed by the branch and bound algorithm.
}
\references{
Eichfelder, Gabriele, Hotz, Thomas, Wieditz, Johannes (2018) - An algorithm for computing Fréchet means on the sphere.
}
\author{
Johannes Wieditz
}
