% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_S2.R
\name{init.acc.sphere}
\alias{init.acc.sphere}
\title{Initialise a list with set entries and the exact lower bound.}
\usage{
init.acc.sphere(set, ...)
}
\arguments{
\item{set}{A list of elements to initialise.}
}
\description{
Initialise a list with set entries and the exact lower bound.
}
\author{
Johannes Wieditz
}
