% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_S2.R
\name{is.in.Delta.qr}
\alias{is.in.Delta.qr}
\title{Check whether a point is contained in a triangle by solving a linear equation.}
\usage{
is.in.Delta.qr(Q, R, y)
}
\arguments{
\item{Q}{The Q matrix of a QR decomposition of the matrix D = (d1, d2, d3) induced by the vertices of the triangle.}

\item{R}{The corresponding R matrix.}

\item{y}{The point to be checked whether it is inside D.}
}
\description{
Check whether a point is contained in a triangle by solving a linear equation.
}
\author{
Johannes Wieditz
}
