% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_S2.R
\name{is.in.Delta.scalprods}
\alias{is.in.Delta.scalprods}
\title{Check whether a point is contained in a triangle by computing scalar products.}
\usage{
is.in.Delta.scalprods(n12, n23, n31, y)
}
\arguments{
\item{n12}{@param n23 @param n31 The normal vectors induced by the vertices of the triangle D.}

\item{y}{The point to be checked whether it is inside D.}
}
\description{
Check whether a point is contained in a triangle by computing scalar products.
}
\author{
Johannes Wieditz
}
