% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_S2.R
\name{is.in.Lambda.qr}
\alias{is.in.Lambda.qr}
\title{Check whether a point is contained in a spherical lune.}
\usage{
is.in.Lambda.qr(Q, R, y)
}
\arguments{
\item{Q}{The Q matrix of a QR decomposition of the matrix Lambda = Lambda(d1, d2) induced by the vertices of the spherical lune.}

\item{R}{The corresponding R matrix.}

\item{y}{The point to be checked whether it is inside Lambda.}
}
\description{
Check whether a point is contained in a spherical lune.
}
\author{
Johannes Wieditz
}
