% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_S1.R
\name{mid.circle}
\alias{mid.circle}
\title{mid.circle computes the midpoint of an intervall}
\usage{
mid.circle(interval)
}
\arguments{
\item{interval}{An interval, i.e. a list with the attributes left and right.}
}
\description{
mid.circle computes the midpoint of an intervall
}
\author{
Johannes Wieditz
}
