% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_S2.R
\name{mid.sphere}
\alias{mid.sphere}
\title{Computes the midpoint of a triangle.}
\usage{
mid.sphere(D)
}
\arguments{
\item{D}{triangle}
}
\description{
Computes the midpoint of a triangle.
}
\author{
Johannes Wieditz
}
