% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_S2.R
\name{norm.vec}
\alias{norm.vec}
\title{Normalise a vector.}
\usage{
norm.vec(vector)
}
\arguments{
\item{vector}{A vector.}
}
\value{
The corresponding unit vector for the passed input. If the input is the origin, we return the origin again.
}
\description{
Normalise a vector.
}
\author{
Johannes Wieditz
}
