% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{output}
\alias{output}
\title{Output function for the results on the circle.}
\usage{
output(A)
}
\arguments{
\item{A}{List which has to be output.}
}
\description{
Output function for the results on the circle.
}
\author{
Johannes Wieditz
}
