% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_functions.R
\name{push}
\alias{push}
\title{Pushing an element to a list. The element is added at the end of the list.}
\usage{
push(L, x)
}
\arguments{
\item{L}{A list.}

\item{x}{An element to be pushed to the list.}
}
\value{
The list L with appended x as last element.
}
\description{
Pushing an element to a list. The element is added at the end of the list.
}
\author{
Johannes Wieditz
}
