% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BranchAndBoundCircle.R
\name{SBB.circle}
\alias{SBB.circle}
\title{The spherical branch and bound algorithm for data on the circle.}
\usage{
SBB.circle(x, ...)
}
\arguments{
\item{x}{Data points on the circle given in angular coordinates.}

\item{...}{Parameters which are passed to the BranchAndBound wrapper, see. BranchAndBound.}
}
\value{
A list of intervals representing a (eps, delta)-approximation of the set of Fréchet means.
}
\description{
The spherical branch and bound algorithm for data on the circle.
}
\examples{
\donttest{
  set.seed(42)
  x    <- runif(100, -pi, pi)
  SBBC <- SBB.circle(x)
}
}
\author{
Johannes Wieditz
}
