% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BranchAndBoundSphere.R
\name{SBB.sphere}
\alias{SBB.sphere}
\title{The spherical branch and bound algorithm for data on the 2-sphere.}
\usage{
SBB.sphere(x, ...)
}
\arguments{
\item{x}{Data points on the circle given in angular coordinates.}

\item{...}{Parameters which are passed to the BranchAndBound wrapper, see. BranchAndBound.}
}
\value{
A list of intervalls representing a (eps, delta)-approximation of the set of Fréchet means.
}
\description{
The spherical branch and bound algorithm for data on the 2-sphere.
}
\examples{

  # Simulated data.
  set.seed(42)
  x <- list()
  for( k in 1:10 ){

     x <- c(x, list(norm.vec(rnorm(3))))
  }
  SBBS <- SBB.sphere(x)
  draw.eps.delta.approximation(SBBS, x)

\donttest{

 # Data from real application.
 require(CircNNTSR)
 data("DataB5FisherSpherical")
 DataB5FisherSpherical <- DataB5FisherSpherical / 180 * pi

 x <- list()
 for( i in 1:length(DataB5FisherSpherical$V1) ){

   x <- c( x , list( c( sin(DataB5FisherSpherical$V1[[i]]) * cos(DataB5FisherSpherical$V2[[i]]), sin(DataB5FisherSpherical$V1[[i]]) * sin(DataB5FisherSpherical$V2[[i]]), cos(DataB5FisherSpherical$V1[[i]]) ) ) )
 }
 SBBS.application <- SBB.sphere(x)
 draw.eps.delta.approximation(SBBS.application, x)

}

}
\author{
Johannes Wieditz
}
