% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_S1.R
\name{size.circle}
\alias{size.circle}
\title{Computes the surface measure of a list of intervals.}
\usage{
size.circle(A)
}
\arguments{
\item{A}{A list of intervals.}
}
\description{
Computes the surface measure of a list of intervals.
}
\author{
Johannes Wieditz
}
