% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_S1.R
\name{size.interval}
\alias{size.interval}
\title{Compute the size of an interval.}
\usage{
size.interval(interval)
}
\arguments{
\item{interval}{An interval (list with entries right and left).}
}
\description{
Compute the size of an interval.
}
\author{
Johannes Wieditz
}
