% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_S2.R
\name{size.sphere}
\alias{size.sphere}
\title{Computes the surface measure of a list of spherical triangles relative to the whole surface.}
\usage{
size.sphere(A)
}
\arguments{
\item{A}{A list of spherical triangles.}
}
\description{
Computes the surface measure of a list of spherical triangles relative to the whole surface.
}
\author{
Johannes Wieditz
}
