% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_S2.R
\name{size.triangle}
\alias{size.triangle}
\title{Compute the circumradius of a triangle.}
\usage{
size.triangle(D)
}
\arguments{
\item{D}{A spherical triangle.}
}
\value{
The circumradius of the (plane) triangle induced by the vertices of D.
}
\description{
Compute the circumradius of a triangle.
}
\author{
Johannes Wieditz
}
