% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BranchAndBoundSphere.R
\name{SBB.sphere}
\alias{SBB.sphere}
\title{The spherical branch and bound algorithm for data on the 2-sphere.}
\usage{
SBB.sphere(x, ...)
}
\arguments{
\item{x}{Data points on the circle given in angular coordinates.}

\item{...}{Parameters which are passed to the BranchAndBound wrapper, see. BranchAndBound.}
}
\value{
A list of intervalls representing a (eps, delta)-approximation of the set of Fréchet means.
}
\description{
The spherical branch and bound algorithm for data on the 2-sphere.
}
\examples{
\donttest{
  set.seed(42)
  x <- list()
  for( k in 1:10 ){

     x <- c(x, list(norm.vec(rnorm(3))))
  }
  SBBS <- SBB.sphere(x)
}

}
\author{
Johannes Wieditz
}
